Удалил комментарии в файлах
